pyproject.toml
[tool.poetry]
name = "smart-iot"
version = "0.1.0"
description = ""
authors = ["Dude <dude@indemeterkast.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django-jazzmin = "^3.0.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
